lint fix.
git-svn-id: file:///svn/dnssec-trigger/trunk@480 14dc9c71-5cc2-e011-b339-0019d10b89f4

+++ riggerd/cfg.c
@@ -247,7 +247,7 @@ static int read_hash(char* arg, struct ssllist* e)
 		/* skip ':' (unless at EOS) */
 		if(*at && i!=len-1) at ++;
 	}
-	hashlist_prepend(&e->hashes, hash, len);
+	hashlist_prepend(&e->hashes, hash, (unsigned int)len);
 	return 1;
 }
